+++ schema/field.go
@@ -5,6 +5,7 @@ import (
 	"reflect"
 
 	"github.com/uptrace/bun/dialect"
+	"github.com/uptrace/bun/internal"
 	"github.com/uptrace/bun/internal/tagparser"
 )
 
@@ -43,6 +44,15 @@ func (f *Field) String() string {
 	return f.Name
 }
 
+func (f *Field) WithIndex(path []int) *Field {
+	if len(path) == 0 {
+		return f
+	}
+	clone := *f
+	clone.Index = makeIndex(path, f.Index)
+	return &clone
+}
+
 func (f *Field) Clone() *Field {
 	cp := *f
 	cp.Index = cp.Index[:len(f.Index):len(f.Index)]
@@ -50,7 +60,7 @@ func (f *Field) Clone() *Field {
 }
 
 func (f *Field) Value(strct reflect.Value) reflect.Value {
-	return fieldByIndexAlloc(strct, f.Index)
+	return internal.FieldByIndexAlloc(strct, f.Index)
 }
 
 func (f *Field) HasNilValue(v reflect.Value) bool {
@@ -102,13 +112,6 @@ func (f *Field) AppendValue(fmter Formatter, b []byte, strct reflect.Value) []by
 	return f.Append(fmter, b, fv)
 }
 
-func (f *Field) ScanWithCheck(fv reflect.Value, src interface{}) error {
-	if f.Scan == nil {
-		return fmt.Errorf("bun: Scan(unsupported %s)", f.IndirectType)
-	}
-	return f.Scan(fv, src)
-}
-
 func (f *Field) ScanValue(strct reflect.Value, src interface{}) error {
 	if src == nil {
 		if fv, ok := fieldByIndex(strct, f.Index); ok {
@@ -117,22 +120,17 @@ func (f *Field) ScanValue(strct reflect.Value, src interface{}) error {
 		return nil
 	}
 
-	fv := fieldByIndexAlloc(strct, f.Index)
+	fv := internal.FieldByIndexAlloc(strct, f.Index)
 	return f.ScanWithCheck(fv, src)
+}
+
+func (f *Field) ScanWithCheck(fv reflect.Value, src interface{}) error {
+	if f.Scan == nil {
+		return fmt.Errorf("bun: Scan(unsupported %s)", f.IndirectType)
+	}
+	return f.Scan(fv, src)
 }
 
 func (f *Field) SkipUpdate() bool {
 	return f.Tag.HasOption("skipupdate")
-}
-
-func indexEqual(ind1, ind2 []int) bool {
-	if len(ind1) != len(ind2) {
-		return false
-	}
-	for i, ind := range ind1 {
-		if ind != ind2[i] {
-			return false
-		}
-	}
-	return true
 }